refactor(admin): extract StatusSection (UI only)
Move status tab content to dedicated component including CourtStatusGrid
and inline waitlist display. Remove unused imports from App.jsx
(icons moved to TabNavigation, CourtStatusGrid to StatusSection).
